Privatize work_executor.hpp
