+++ tt_metal/api/tt-metalium/device.hpp
@@ -9,7 +9,7 @@
 #include <utility>
 
 #include "hostdevcommon/common_values.hpp"
-#include "work_executor.hpp"
+#include "work_executor_types.hpp"
 #include "basic_allocator.hpp"
 #include "l1_banking_allocator.hpp"
 #include "data_types.hpp"
