+++ tt_metal/common/CMakeLists.txt
@@ -20,6 +20,7 @@ target_include_directories(
     PUBLIC
         ${PROJECT_SOURCE_DIR}
         ${PROJECT_SOURCE_DIR}/tt_metal
+        ${CMAKE_CURRENT_SOURCE_DIR}
 )
 
 target_link_libraries(
