+++ tt_metal/impl/dispatch/hardware_command_queue.cpp
@@ -14,6 +14,8 @@
 
 #include "tt_cluster.hpp"
 
+#include "work_executor.hpp"
+
 // Because we are a Friend of Program, accessing Program::get_program_transfer_info() and Program::get_kernels_buffer()
 // MUST REMOVE
 #include <program_impl.hpp>
